# Speed up tests by reducing kernel loads in test fixtures.