--- a/mujoco_warp/_src/collision_driver_test.py
+++ b/mujoco_warp/_src/collision_driver_test.py
@@ -295,7 +295,8 @@ class CollisionTest(parameterized.TestCase):
     # generate contact
     m = mjwarp.put_model(mjm)
     d = mjwarp.make_data(mjm)
-    mjwarp.forward(m, d)
+    mjwarp.kinematics(m, d)
+    mjwarp.collision(m, d)
 
     self.assertEqual(d.ncon.numpy()[0], 1)
     self.assertEqual(d.contact.includemargin.numpy()[0], -1)
@@ -333,7 +334,8 @@ class CollisionTest(parameterized.TestCase):
     # generate contact
     m = mjwarp.put_model(mjm)
     d = mjwarp.make_data(mjm)
-    mjwarp.forward(m, d)
+    mjwarp.kinematics(m, d)
+    mjwarp.collision(m, d)
 
     self.assertEqual(d.ncon.numpy()[0], 1)
     self.assertEqual(d.contact.includemargin.numpy()[0], -1)
@@ -372,7 +374,8 @@ class CollisionTest(parameterized.TestCase):
     # generate contact
     m = mjwarp.put_model(mjm)
     d = mjwarp.make_data(mjm)
-    mjwarp.forward(m, d)
+    mjwarp.kinematics(m, d)
+    mjwarp.collision(m, d)
 
     self.assertEqual(d.ncon.numpy()[0], 1)
     self.assertEqual(d.contact.includemargin.numpy()[0], -1)
@@ -415,7 +418,8 @@ class CollisionTest(parameterized.TestCase):
     # generate contact
     m = mjwarp.put_model(mjm)
     d = mjwarp.make_data(mjm)
-    mjwarp.forward(m, d)
+    mjwarp.kinematics(m, d)
+    mjwarp.collision(m, d)
 
     self.assertEqual(d.ncon.numpy()[0], 2)
     self.assertEqual(d.contact.includemargin.numpy()[1], -1)