--- a/mujoco_warp/test_data/actuation/actuators.xml
+++ b/mujoco_warp/test_data/actuation/actuators.xml
@@ -1,4 +1,5 @@
 <mujoco>
+  <option iterations="4" ls_iterations="4"/>
   <worldbody>
     <body>
       <joint name="hinge" type="hinge"/>