Currently, default `iterations` and `ls_iterations` is 50 and 50.  Generally this is OK, but one implication is that kernels inside linesearch are loaded 50 x 50 = 2500 times, which even when cached can be slow to do for a physics step.

Probably after #78 lands we should revisit what our `iterations` and `ls_iterations` defaults should be, whether we should turn on parallel linesearch by default, or whether there's some other way to reduce the cost of 2500 kernel cache hits.

For the time being this PR removes the places in our tests where we were hitting these large kernel cache hits, it lowers test times on my machine from 82s (after cache is fully populated) to 11s.

Output of `pytest --durations=10` before this PR:

```
========================================================================================================= slowest 10 durations ==========================================================================================================
34.68s call     mujoco_warp/_src/solver_test.py::SolverTest::test_frictionloss
18.44s call     mujoco_warp/_src/collision_driver_test.py::CollisionTest::test_contact_pair
13.53s call     mujoco_warp/_src/forward_test.py::ForwardTest::test_rungekutta4
4.81s call     mujoco_warp/_src/forward_test.py::ForwardTest::test_actuation1
3.56s call     mujoco_warp/_src/solver_test.py::SolverTest::test_solve0
1.19s call     mujoco_warp/_src/solver_test.py::SolverTest::test_solve_batch0
0.46s call     mujoco_warp/_src/solver_test.py::SolverTest::test_solve1
0.42s call     mujoco_warp/_src/broad_phase_test.py::BroadphaseTest::test_nxn_broadphase
0.41s call     mujoco_warp/_src/solver_test.py::SolverTest::test_solve2
0.40s call     mujoco_warp/_src/collision_driver_test.py::CollisionTest::test_collision_disableflags0
==================================================================================================== 118 passed in 82.17s (0:01:22) =====================================================================================================
```

And after this PR:

```
========================================================================================================= slowest 10 durations ==========================================================================================================
1.04s call     mujoco_warp/_src/solver_test.py::SolverTest::test_solve_batch0
0.84s call     mujoco_warp/_src/solver_test.py::SolverTest::test_frictionloss0
0.84s call     mujoco_warp/_src/solver_test.py::SolverTest::test_solve0
0.79s call     mujoco_warp/_src/solver_test.py::SolverTest::test_frictionloss1
0.76s call     mujoco_warp/_src/forward_test.py::ForwardTest::test_rungekutta4
0.52s call     mujoco_warp/_src/solver_test.py::SolverTest::test_solve2
0.52s call     mujoco_warp/_src/solver_test.py::SolverTest::test_solve1
0.46s call     mujoco_warp/_src/forward_test.py::ImplicitIntegratorTest::test_implicit0
0.41s call     mujoco_warp/_src/broad_phase_test.py::BroadphaseTest::test_nxn_broadphase
0.40s call     mujoco_warp/_src/forward_test.py::ForwardTest::test_actuation1
========================================================================================================= 119 passed in 11.03s ==========================================================================================================
```