@yuvaltassa 2025-04-17
My expectation is that grid-refinement parallel linesearch will be great for CG and less great for Newton steps. 

While I don't think we should expose this option in XML, I'd like to have it in mjOption and implement on CPU (toggleable only procedurally, since it never makes sense on 1 thread). I expect there is a better algorithm than sequential grid refinement that looks like a bunch of partially overlapping brackets with quadratic/cubic fits per bracket. Do you see it?

Another thought I'm having is `int mjOption.iterations_solver[3]`, initialized to {0, 0, 0}, that overrides .iterations (per solver) if positive.